Implement test block device interface in userspace
Pull get_block_device_from_name() into getIndexTestNames.c and inline
userspace methods to simplify the code. Use this method to implement
the new test block device methods in userspace.

Remove inlined methods and methods which are no longer used. This
includes removing the entirety of blkdev.c.
