+++ src/c++/uds/userLinux/uds/linux/blkdev.h
@@ -11,7 +11,6 @@
 
 #include <linux/compiler_attributes.h>
 #include <linux/types.h>
-#include <linux/version.h>
 #include <stdio.h>
 
 #define SECTOR_SHIFT    9
@@ -24,15 +23,9 @@
 
 #define MAJOR(dev) ((unsigned int) ((dev) >> MINORBITS))
 #define MINOR(dev) ((unsigned int) ((dev) & MINORMASK))
-#define MKDEV(ma,mi) (((ma) << MINORBITS) | (mi))
 
 #define format_dev_t(buffer, dev)				\
 	sprintf(buffer, "%u:%u", MAJOR(dev), MINOR(dev))
-
-#undef VDO_USE_ALTERNATE
-#if (LINUX_VERSION_CODE < KERNEL_VERSION(6, 5, 0))
-#define VDO_USE_ALTERNATE
-#endif
 
 /* Defined in linux/fs.h but hacked for vdo unit testing */
 struct inode {
@@ -53,12 +46,6 @@ struct block_device {
 	struct inode *bd_inode;
 };
 
-#ifndef VDO_USE_ALTERNATE
-struct blk_holder_ops {
-	void (*mark_dead)(struct block_device *bdev);
-};
-
-#endif /* !VDO_USE_ALTERNATE */
 /**********************************************************************/
 static inline int blk_status_to_errno(blk_status_t status)
 {
@@ -73,43 +60,6 @@ static inline blk_status_t errno_to_blk_status(int error)
 
 /**********************************************************************/
 blk_qc_t submit_bio_noacct(struct bio *bio);
-
-/**********************************************************************/
-#ifdef VDO_USE_ALTERNATE
-static inline struct block_device *
-blkdev_get_by_dev(dev_t dev __always_unused,
-		  fmode_t mode __always_unused,
-		  void *holder __always_unused)
-#else
-static inline struct block_device *
-blkdev_get_by_dev(dev_t dev __always_unused,
-		  fmode_t mode __always_unused,
-		  void *holder __always_unused,
-		  const struct blk_holder_ops *hops __always_unused)
-#endif /* VDO_USE_ALTERNATE */
-{
-	/* This function will not get called in userspace. */
-	return 0;
-}
-
-/**********************************************************************/
-#ifdef VDO_USE_ALTERNATE
-struct block_device *blkdev_get_by_path(const char *path,
-					fmode_t mode,
-					void *holder);
-#else
-struct block_device *blkdev_get_by_path(const char *path,
-					fmode_t mode,
-					void *holder,
-					const struct blk_holder_ops *hops __always_unused);
-#endif /* VDO_USE_ALTERNATE */
-
-/**********************************************************************/
-#ifdef VDO_USE_ALTERNATE
-void blkdev_put(struct block_device *bdev, fmode_t mode);
-#else
-void blkdev_put(struct block_device *bdev, void *holder);
-#endif /* VDO_USE_ALTERNATE */
 
 /* Defined in linux/fs.h, but it's convenient to implement here. */
 static inline loff_t i_size_read(const struct inode *inode)
